添加请求错误的异常判断逻辑

+++ okhttputils/src/main/java/kz/ally/okhttp/client/AbsResponse.java
@@ -6,7 +6,7 @@ package kz.ally.okhttp.client;
  */
 public abstract class AbsResponse {
 
-    public long code;
+    public int code;
 
     public String message;
